refactor: 투표한 트랙에 관한 정보를 VoteResultTracks 컴포넌트로 분리
